[FIX] Clean up and test page id / TS generation in BE for TYPO3 7.x and 8.x

+++ Classes/Configuration/BackendConfigurationManager.php
@@ -24,7 +24,10 @@ class BackendConfigurationManager extends \TYPO3\CMS\Extbase\Configuration\Backe
      * Make sure TS is generated from currently selected page
      *
      * Needed basically to make extbase work. Without PID is set to 0 (which is root)
-     * and persistence, TS generation, etc. will fail.
+     * and persistence, TS generation, etc. will fail. This is the case in TYPO3 8.x
+     * when editing a record using the context menu (t3-js-clickmenutrigger CSS class).
+     *
+     * Using a issueCommand VH link works as expected. Not needed for TYPO3 7.x.
      *
      * @todo Rework this: ugly hack but not sure how to solve this in a clean way
      *
@@ -34,10 +37,10 @@ class BackendConfigurationManager extends \TYPO3\CMS\Extbase\Configuration\Backe
         $id = parent::getCurrentPageIdFromGetPostData();
 
         if (empty($id)){
-            $id = GeneralUtility::_GP('popViewId');
+            $id = (int) GeneralUtility::_GP('popViewId');
         }
 
-        return (int) $id;
+        return $id;
     }
 
 }
